pyproject.toml
[project]
name = "zp-simulation"
version = "0.1.0"
description = "Segmentation pipeline for pronuclei and blastocyst analysis"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Deep Learning Framework
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    
    # Segmentation Models
    "segmentation-models-pytorch>=0.3.3",
    
    # Image Processing and Computer Vision
    "opencv-python>=4.8.0",
    "Pillow>=10.0.0",
    "scikit-image>=0.21.0",
    "albumentations>=1.3.1",
    
    # Numerical and Scientific Computing
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    
    # Visualization and Monitoring
    "matplotlib>=3.7.0",
    "tensorboard>=2.13.0",
    
    # Utilities
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "flake8>=6.1.0",
    "isort>=5.12.0",
    "mypy>=1.4.0",
]

notebook = [
    "jupyter>=1.0.0",
    "ipykernel>=6.25.0",
    "notebook>=7.0.0",
]

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
